+++ cypress/support/commands.ts
@@ -26,8 +26,6 @@ import {
   mockGetLoginSchemaType,
   mockGetMemberBy,
   mockGetMembers,
-  mockGetOwnItems,
-  mockGetSharedItems,
   mockPatchAppData,
   mockPostAppData,
   mockProfilePage,
@@ -38,6 +36,7 @@ Cypress.Commands.add(
   'setUpApi',
   ({
     items = [],
+    itemLogins = [],
     chatMessages = [],
     members = Object.values(MEMBERS),
     currentMember = CURRENT_USER,
@@ -47,11 +46,8 @@ Cypress.Commands.add(
     getCurrentMemberError = false,
   } = {}) => {
     if (currentMember) {
-      cy.setCookie(CookieKeys.Session, 'somecookie');
       cy.setCookie(CookieKeys.AcceptCookies, 'true');
     }
-    mockGetOwnItems({ items, currentMember });
-    mockGetSharedItems({ items, currentMember });
     mockGetAccessibleItems(items);
     mockGetItem(
       { items, currentMember },
@@ -63,7 +59,7 @@ Cypress.Commands.add(
     mockGetItemTags(items, currentMember);
 
     mockGetItemsTags(items, currentMember);
-    mockGetLoginSchemaType(items, currentMember);
+    mockGetLoginSchemaType(itemLogins, currentMember);
 
     mockGetChildren(items, currentMember);
 
@@ -125,6 +121,7 @@ declare global {
     interface Chainable {
       setUpApi({
         items,
+        itemLogins,
         members,
         chatMessages,
         currentMember,
@@ -135,6 +132,7 @@ declare global {
         getAppLinkError,
       }?: {
         items?: MockItem[];
+        itemLogins?: { [key: string]: string };
         members?: Member[];
         chatMessages?: ChatMessage[];
         currentMember?: Member | null;