+++ src/contexts/CurrentMemberContext.tsx
@@ -1,5 +1,8 @@
 import { createContext, useContext, useEffect } from 'react';
 
+import { AccountType } from '@graasp/sdk';
+import { DEFAULT_LANG } from '@graasp/translations';
+
 import i18n from '@/config/i18n';
 import { hooks } from '@/config/queryClient';
 
@@ -20,7 +23,10 @@ export const CurrentMemberContextProvider = ({
   const query = useCurrentMember();
 
   // update language depending on user setting
-  const lang = query.data?.extra?.lang;
+  const lang =
+    query.data && query.data?.type === AccountType.Individual
+      ? query.data?.extra?.lang
+      : DEFAULT_LANG;
   useEffect(() => {
     if (lang !== i18n.language) {
       i18n.changeLanguage(lang);