+++ src/modules/userSwitch/MemberAvatar.tsx
@@ -1,4 +1,4 @@
-import { Member, ThumbnailSize } from '@graasp/sdk';
+import { CurrentAccount, ThumbnailSize } from '@graasp/sdk';
 import { COMMON } from '@graasp/translations';
 import { Avatar } from '@graasp/ui';
 
@@ -8,7 +8,7 @@ import { hooks } from '@/config/queryClient';
 import { buildMemberAvatarId } from '@/config/selectors';
 
 type Props = {
-  member?: Member | null;
+  member?: CurrentAccount | null;
 };
 
 const MemberAvatar = ({ member }: Props): JSX.Element => {